chore(types): organize types
