+++ src/pages/announcements/[slug].tsx
@@ -70,12 +70,10 @@ const AnnouncementPage: NextPage<ArticlePageProps> = ({
   ) : (
     <ArticleIndex
       breadcrumbList={breadcrumbList}
-      name={componentProps?.articleData?.name ?? ''}
-      children={componentProps?.articleData?.children}
-      hidePaginationNext={componentProps?.articleData?.hidePaginationNext}
-      hidePaginationPrevious={
-        componentProps?.articleData?.hidePaginationPrevious
-      }
+      name={componentProps?.name ?? ''}
+      children={componentProps?.children}
+      hidePaginationNext={componentProps?.hidePaginationNext}
+      hidePaginationPrevious={componentProps?.hidePaginationPrevious}
       isListed={isListed}
       slug={slug}
       pagination={pagination}
@@ -232,12 +230,10 @@ export const getStaticProps: GetStaticProps = async ({
           breadcrumbList,
           branch,
           componentProps: {
-            articleData: {
-              name: categoryTitle || slug,
-              children: childrenList,
-              hidePaginationPrevious: breadcrumbList.length < 2,
-              hidePaginationNext: !childrenList.length,
-            },
+            name: categoryTitle || slug,
+            children: childrenList,
+            hidePaginationPrevious: breadcrumbList.length < 2,
+            hidePaginationNext: !childrenList.length,
           },
           locale: currentLocale,
         },
