+++ src/pages/docs/tutorials/[slug].tsx
@@ -83,12 +83,10 @@ const TutorialPage: NextPage<ArticlePageProps> = ({
   ) : (
     <ArticleIndexing
       breadcrumbList={breadcrumbList}
-      name={componentProps?.articleData?.name ?? ''}
-      children={componentProps?.articleData?.children}
-      hidePaginationNext={componentProps?.articleData?.hidePaginationNext}
-      hidePaginationPrevious={
-        componentProps?.articleData?.hidePaginationPrevious
-      }
+      name={componentProps?.name ?? ''}
+      children={componentProps?.children}
+      hidePaginationNext={componentProps?.hidePaginationNext}
+      hidePaginationPrevious={componentProps?.hidePaginationPrevious}
       isListed={isListed}
       slug={slug}
       pagination={pagination}
@@ -269,12 +267,10 @@ export const getStaticProps: GetStaticProps = async ({
           breadcrumbList,
           branch,
           componentProps: {
-            articleData: {
-              name: categoryTitle || slug,
-              children: childrenList,
-              hidePaginationPrevious: breadcrumbList.length < 2,
-              hidePaginationNext: !childrenList.length,
-            },
+            name: categoryTitle || slug,
+            children: childrenList,
+            hidePaginationPrevious: breadcrumbList.length < 2,
+            hidePaginationNext: !childrenList.length,
           },
           locale: currentLocale,
         },
