+++ src/pages/known-issues/[slug].tsx
@@ -71,12 +71,10 @@ const KnownIssuePage: NextPage<ArticlePageProps> = ({
   ) : (
     <ArticleIndex
       breadcrumbList={breadcrumbList}
-      name={componentProps?.articleData?.name ?? ''}
-      children={componentProps?.articleData?.children}
-      hidePaginationNext={componentProps?.articleData?.hidePaginationNext}
-      hidePaginationPrevious={
-        componentProps?.articleData?.hidePaginationPrevious
-      }
+      name={componentProps?.name ?? ''}
+      children={componentProps?.children}
+      hidePaginationNext={componentProps?.hidePaginationNext}
+      hidePaginationPrevious={componentProps?.hidePaginationPrevious}
       isListed={isListed}
       slug={slug}
       pagination={pagination}
@@ -240,11 +238,9 @@ export const getStaticProps: GetStaticProps = async ({
         breadcrumbList,
         branch,
         componentProps: {
-          articleData: {
-            name: categoryTitle || slug,
-            children: childrenList,
-            hidePaginationNext: !childrenList.length,
-          },
+          name: categoryTitle || slug,
+          children: childrenList,
+          hidePaginationNext: !childrenList.length,
         },
         locale: currentLocale,
       },
