+++ src/pages/troubleshooting/[slug].tsx
@@ -67,12 +67,10 @@ const TroubleshootingPage: NextPage<ArticlePageProps> = ({
   ) : (
     <ArticleIndex
       breadcrumbList={breadcrumbList}
-      name={componentProps?.articleData?.name ?? ''}
-      children={componentProps?.articleData?.children}
-      hidePaginationNext={componentProps?.articleData?.hidePaginationNext}
-      hidePaginationPrevious={
-        componentProps?.articleData?.hidePaginationPrevious
-      }
+      name={componentProps?.name ?? ''}
+      children={componentProps?.children}
+      hidePaginationNext={componentProps?.hidePaginationNext}
+      hidePaginationPrevious={componentProps?.hidePaginationPrevious}
       isListed={isListed}
       slug={slug}
       pagination={pagination}
@@ -218,11 +216,9 @@ export const getStaticProps: GetStaticProps = async ({
         breadcrumbList,
         branch,
         componentProps: {
-          articleData: {
-            name: categoryTitle || slug,
-            children: childrenList,
-            hidePaginationNext: !childrenList.length,
-          },
+          name: categoryTitle || slug,
+          children: childrenList,
+          hidePaginationNext: !childrenList.length,
         },
         locale: currentLocale,
       },
